Refactor asciidox generator into subpackage
